Remove all association from this project, please read commit description
Due to disagreements with the current Salhack moderation team, I'd like to remove all association I have with this project, especially because of how I was treated after making an announcement trying to help the current users of a dead project, along with offering to maintain this project. 

I do not wish to maintain this project after these events, and I'd like to have my push access removed.